services/gateway/requirements.dev.txt
anyio==3.7.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
exceptiongroup==1.1.1
fastapi==0.96.0
filelock==3.12.0
grpcio==1.54.2
grpcio-tools==1.54.2
h11==0.14.0
identify==2.5.24
idna==3.4
nodeenv==1.8.0
platformdirs==3.5.1
pre-commit==3.3.2
protobuf==4.23.2
pydantic==1.10.9
python-decouple==3.8
python-multipart==0.0.6
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
virtualenv==20.23.0